Game Logic + Refactor
All right here we go.
Created new stat system; there are 3 hard stats: Strength, Dexterity, and Intellect. These are used to calculate the values of 3 soft stats: Vitality, Endurance, and Willpower. These soft stats are then used to calculate the MAXIMUM value of 3 consumable stats: Health, Stamina, and Will.
The 3 hard stats are in the stat block object and are stored with the 3 consumable stats' max value. Each creature has its own map of the 3 consumable stats which gets depleted by the damage() and heal() functions.
I have used EnumMap for the stat blocks because it is faster than HashMap for this usage.
Refactored Player, Creature, and Mob entities to use new stat system. It is not possible to alter a creature's maximum consumable stat directly, this must be done by altering their hard stats.